+++ lib/empty_eye/active_record/base.rb
@@ -1,11 +1,7 @@
-module EmptyEye
-  module Base
+module ActiveRecord
+  class Base
     
-    def self.included(base)
-      base.extend ClassMethods
-    end
-    
-    module ClassMethods
+    class << self
     
       def mti_class?
         extended_with.any?
@@ -13,9 +9,9 @@ module EmptyEye
     
       def mti_class(primary_table = nil)
         self.primary_key = "id"
-        raise(EmptyEye::AlreadyExtended, "extend table method already invoked") if mti_class?
+        raise(EmptyEye::AlreadyExtended, "MTI class method already invoked") if mti_class?
         set_mti_primary_table(primary_table)
-        self.table_name = compute_table_name 
+        self.table_name = compute_view_name
         extended_with.primary_table(mti_primary_table)
         before_yield = reflect_on_multiple_associations(:has_one)
         yield nil if block_given?
@@ -29,6 +25,8 @@ module EmptyEye
           extended_with.association(assoc)
         end
         create_view
+        reset_column_information
+        inherit_mti_validations
       end
       
       def mti_associations
@@ -36,18 +34,32 @@ module EmptyEye
       end
 
       def extended_with
-        @extended_with ||= superclass_extensions || ViewExtensionCollection.new(self)
+        @extended_with ||= EmptyEye::ViewExtensionCollection.new(self)
       end
 
-      def primary_shard
+      def mti_primary_shard
         extended_with.primary.shard
       end
       
+      def finder_needs_type_condition?
+        !mti_class? and super
+      end
+
+      def mti_shard?
+        false
+      end
+      
+      private
+      
+      def compute_view_name
+        descends_from_active_record? ? compute_table_name : name.underscore.pluralize
+      end
+      
       def set_mti_primary_table(primary_table_name)
-        @mti_primary_table = if primary_table_name.nil?
+        @mti_primary_table = if ordinary_table_exists?
+          raise(EmptyEye::ViewNameError, "MTI view cannot be created because a table named '#{compute_view_name}' already exists")
+        elsif primary_table_name.nil?
           descends_from_active_record? ? "#{compute_table_name}_core" : superclass.table_name
-        elsif ordinary_table_exists?
-          raise(EmptyEye::ViewNameError, "a table named '#{primary_table_name}' already exists")
         else
           primary_table_name
         end
@@ -64,7 +76,7 @@ module EmptyEye
       end
 
       def ordinary_table_exists?
-        connection.tables_without_views.include?(compute_table_name)
+        connection.tables_without_views.include?(compute_view_name)
       end
     
       def create_view
@@ -73,22 +85,36 @@ module EmptyEye
       end
     
       def superclass_extensions
-        superclass.extended_with.dup unless descends_from_active_record?
+        superclass.extended_with.dup.descend(self) unless descends_from_active_record?
+      end
+      
+      def inherit_mti_validations
+        extended_with.validations.each {|args| send(*args)}
+      end
+    end
+    
+    def mti_shard_instance?
+      false
+    end
+    
+    def base_class_name
+      if mti_shard_instance?
+        mti_master_class.base_class.name
+      else
+        self.class.base_class.name
       end
     end
     
     private
     
-    def primary_shard
-      @primary_shard ||= if new_record?
-        self.class.primary_shard.new(:mti_instance => self)
+    def mti_primary_shard
+      @mti_primary_shard ||= if new_record?
+        self.class.mti_primary_shard.new(:mti_instance => self)
       else
-        self.class.primary_shard.find_by_id(id)
+        rtn = self.class.mti_primary_shard.find_by_id(id)
+        rtn.mti_instance = self
+        rtn
       end
-    end
-    
-    def table_extended_with
-      self.class.extended_with
     end
     
     def mti_class?