+++ lib/empty_eye/associations/builder/shard_has_one.rb
@@ -1,0 +1,15 @@
+module EmptyEye
+  module Associations
+    module Builder
+      class ShardHasOne < ActiveRecord::Associations::Builder::HasOne
+        
+        def build
+          reflection = super
+          configure_dependency unless options[:through]
+          reflection
+        end
+        
+      end
+    end
+  end
+end