+++ spec/mti_crud_spec.rb
@@ -5,7 +5,6 @@ require 'bundler/setup'
 describe ActiveRecord::Base do
   before(:each) do
     #these classes where added for testing convenience
-    RestaurantCore.delete_all
     BarCore.delete_all
     
     Business.delete_all
@@ -59,19 +58,75 @@ describe ActiveRecord::Base do
       @bar.phone.should eq("987654321")
       @bar.business.phone.should eq("987654321")
     end
+    
+    it "should update a mti class correctly with Class.update" do
+      Bar.update(@bar.id, :name => 'Betos')
+      @bar.reload
+      @bar.name.should eq('Betos')
+    end
+    
+    it "should update a mti class correctly with Class.update_all" do
+      rtn = Bar.update_all(:name => 'Betos')
+      @bar.reload
+      rtn.should eq(1)
+      @bar.name.should eq('Betos')
+    end
+    
+    it "should not update a mti class incorrectly with Class.update_all" do
+      rtn = Bar.update_all({:name => 'Betos'}, ["id = ?", @bar.id + 1]) #choose the wrong one
+      @bar.reload
+      rtn.should eq(0)
+      @bar.name.should eq('Chicos')
+    end
   end
   
   describe "delete" do
     it "should destroy a mti class correctly" do
+      @business = @bar.business
       @bar.destroy
       @bar.destroyed?.should eq(true)
       Bar.find_by_id(@bar.id).should eq(nil)
+      Business.find_by_id(@business.id).should eq(nil)
+    end
+    
+    it "should destroy_all mti class correctly" do
+      Bar.count.should eq(1)
+      Business.count.should eq(1)
+      Bar.destroy_all
+      Bar.count.should eq(0)
+      Business.count.should eq(0)
+    end
+    
+    it "should not destroy_all mti class incorrectly" do
+      Bar.count.should eq(1)
+      Business.count.should eq(1)
+      Bar.destroy_all(:id => @bar.id + 1) #choose the wrong one
+      Bar.count.should eq(1)
+      Business.count.should eq(1)
     end
     
     it "should delete a mti class correctly" do
       @bar.delete
       @bar.destroyed?.should eq(true)
       Bar.find_by_id(@bar.id).should eq(nil)
+    end
+    
+    it "should delete_all mti class correctly" do
+      Bar.count.should eq(1)
+      Business.count.should eq(1)
+      rtn = Bar.delete_all
+      rtn.should eq(1)
+      Bar.count.should eq(0)
+      Business.count.should eq(0)
+    end
+    
+    it "should not delete_all mti class incorrectly" do
+      Bar.count.should eq(1)
+      Business.count.should eq(1)
+      rtn = Bar.delete_all(:id => @bar.id + 1) #choose the wrong one
+      rtn.should eq(0)
+      Bar.count.should eq(1)
+      Business.count.should eq(1)
     end
   end
 end